+++ src/ingest_validation_tools/upload.py
@@ -4,7 +4,7 @@ import logging
 import subprocess
 from collections import Counter, defaultdict
 from copy import copy
-from dataclasses import dataclass, field
+from dataclasses import dataclass, field, fields
 from datetime import datetime
 from fnmatch import fnmatch
 from functools import cached_property
@@ -33,6 +33,12 @@ from ingest_validation_tools.validation_utils import (
 TSV_SUFFIX = "metadata.tsv"
 
 
+class ErrorDictException(Exception):
+    def __init__(self, error: str):
+        super().__init__(error)
+        self.error = error
+
+
 @dataclass
 class InfoDict:
     time: datetime
@@ -52,44 +58,86 @@ class InfoDict:
 @dataclass
 class ErrorDict:
     preflight: List[str] = field(default_factory=list)
-    directory: Dict[str, List] = field(default_factory=lambda: defaultdict(list))
-    upload_metadata: Dict[str, Dict[str, str]] = field(default_factory=lambda: defaultdict(dict))
-    metadata_validation_local: DefaultDict[str, Dict] = field(
-        default_factory=lambda: defaultdict(dict)
+    directory: DefaultDict[str, List[str]] = field(default_factory=lambda: defaultdict(list))
+    upload_metadata: DefaultDict[str, List[str]] = field(default_factory=lambda: defaultdict(list))
+    metadata_validation_local: DefaultDict[str, List[str]] = field(
+        default_factory=lambda: defaultdict(list)
     )
-    metadata_validation_api: DefaultDict[str, Dict] = field(
-        default_factory=lambda: defaultdict(dict)
+    metadata_validation_api: DefaultDict[str, List] = field(
+        default_factory=lambda: defaultdict(list)
     )
     metadata_url_errors: DefaultDict[str, List] = field(default_factory=lambda: defaultdict(list))
-    reference: Dict[str, List] = field(default_factory=dict)
-    plugin: Dict[str, List] = field(default_factory=dict)
+    reference: Dict[str, List[str]] = field(default_factory=dict)
+    plugin: Dict[str, List[str]] = field(default_factory=dict)
     plugin_skip: Optional[str] = None
 
     def __bool__(self):
         return bool(self.as_dict())
 
+    @property
+    def field_map(self):
+        return {
+            "preflight": "Preflight Errors",
+            "directory": "Directory Errors",
+            "upload_metadata": "Antibodies/Contributors Errors",
+            "metadata_validation_local": "Local Validation Errors",
+            "metadata_validation_api": "API Validation Errors",
+            "metadata_url_errors": "URL Check Errors",
+            "reference": "Reference Errors",
+            "plugin": "Data File Errors",
+            "plugin_skip": "Fatal Errors",
+        }
+
+    @property
+    def metadata_only_fields(self):
+        return [
+            self.metadata_validation_local,
+            self.metadata_url_errors,
+            self.metadata_validation_api,
+        ]
+
+    def metadata_only_errors(self):
+        return {
+            self.field_map.get(error_field.name): self.sort_val(getattr(self, error_field.name))
+            for error_field in self.metadata_only_fields
+            if getattr(self, error_field.name)
+        }
+
+    def tsv_only_errors_by_path(self, path: str, local_allowed=False):
+        """
+        For use in front-end single TSV validation.
+        """
+        errors = {}
+        for key, value in self.metadata_validation_api.items():
+            if Path(key) == Path(path):
+                errors["API Validation Errors"] = self.sort_val(value)
+                break
+        for key, value in self.metadata_url_errors.items():
+            if Path(key) == Path(path):
+                errors["URL Errors"] = self.sort_val(value)
+                break
+        if local_allowed:
+            for key, value in self.metadata_validation_local.items():
+                if Path(key) == Path(path):
+                    errors["Local Validation Errors"] = self.sort_val(value)
+                    break
+        return errors
+
     def as_dict(self):
         errors = {}
-        # self.compile_sections()
-        if self.preflight:
-            errors["Preflight Errors"] = self.preflight
-        if self.directory:
-            errors["Directory Errors"] = self.directory
-        if self.upload_metadata:
-            errors["Antibodies/Contributors TSV Errors"] = self.upload_metadata
-        if self.metadata_validation_local:
-            errors["Local Validation Errors"] = self.metadata_validation_local
-        if self.metadata_validation_api:
-            errors["API Validation Errors"] = self.metadata_validation_api
-        if self.metadata_url_errors:
-            errors["URL Check Errors"] = self.metadata_url_errors
-        if self.reference:
-            errors["Reference Errors"] = self.reference
-        if self.plugin_skip:
-            errors["Fatal Exception"] = self.plugin_skip
-        elif self.plugin:
-            errors["Plugin Errors"] = self.plugin
+        for error_field in fields(self):
+            value = getattr(self, error_field.name)
+            if value:
+                value = self.sort_val(value)
+                errors[self.field_map.get(error_field.name)] = getattr(self, error_field.name)
         return errors
+
+    def sort_val(self, value):
+        if type(value) is list:
+            value = sorted(value)
+        elif type(value) in [dict, defaultdict]:
+            value = {k: value[k] for k in sorted(value.keys())}
+        return value
 
 
 class Upload:
@@ -188,12 +236,14 @@ class Upload:
             stderr=subprocess.STDOUT,
         ).strip()
 
+        # If called before get_errors, will report dir schema major version only
+
         try:
             tsvs = {
                 Path(path).name: {
                     "Schema": sv.table_schema,
                     "Metadata schema version": sv.version,
-                    "Directory schema versions": sv.dir_schema,
+                    "Directory schema version": sv.dir_schema,
                 }
                 for path, sv in self.effective_tsv_paths.items()
             }
@@ -236,7 +286,6 @@ class Upload:
 
         return self.errors
 
-    # Optimization: this exists to support get_tsv_errors and contributors/antibodies validation; simplify path?
     def validation_routine(
         self,
         report_type: ReportType = ReportType.STR,
@@ -267,39 +316,32 @@ class Upload:
     def _get_local_tsv_errors(self):
         for path, schema in self.effective_tsv_paths.items():
             if "data_path" not in schema.rows[0] or "contributors_path" not in schema.rows[0]:
-                self.errors.upload_metadata[path].update(
-                    {
-                        f"{path} (as {schema.table_schema})": "File is missing data_path or contributors_path."
-                    }
+                self.errors.upload_metadata[f"{path} (as {schema.table_schema})"].append(
+                    "File is missing data_path or contributors_path."
                 )
             for ref in ["contributors", "antibodies"]:
                 errors = self._get_ref_errors(ref, schema, path)
                 if errors:
-                    self.errors.upload_metadata[path].update(errors)
+                    self.errors.upload_metadata.update(errors)
 
-    def _get_directory_errors(self) -> dict:
-        errors = {}
+    def _get_directory_errors(self):
         if self.is_multi_assay and self.multi_parent:
             for data_path in self.multi_assay_data_paths:
+                # TODO
+                assert Path(data_path).exists, ""
                 dir_errors = self._check_data_path(
                     self.multi_parent, Path(self.multi_parent.path), data_path
                 )
-                self.errors.directory[data_path].append(dir_errors)
+                for schema in self.effective_tsv_paths.values():
+                    if not schema.dataset_type == self.multi_parent.dataset_type:
+                        schema.dir_schema = self.multi_parent.dir_schema
+                if dir_errors:
+                    self.errors.directory.update(dir_errors)
         else:
             for path, schema in self.effective_tsv_paths.items():
                 dir_errors = self._get_ref_errors("data", schema, path)
                 if dir_errors:
-                    """
-                    TODO: there's an issue here (and any other places setting a key
-                    that might be long) where the YAML dumper converts this to a
-                    complex key and adds a ? at the front and inserts a line break
-                    at the next instance of whitespace (in this case, leading
-                    (as x) to be on a following line)
-                    Wrote validation_utils > print_path to try to address this
-                    issue, but needed to deprioritize.
-                    """
-                    self.errors.directory[path].append(dir_errors)
-        return errors
+                    self.errors.directory.update(dir_errors)
 
     def _validate(
         self,
@@ -319,20 +361,24 @@ class Upload:
                     self.offline,
                 )
             except Exception as e:
-                self.errors.metadata_validation_local[tsv_path].update(
-                    {f"{tsv_path} (as {schema_version.table_schema})": e}
+                self.errors.metadata_validation_local.update(
+                    {f"{tsv_path} (as {schema_version.table_schema})": [str(e)]}
+                )
+                return
+            if schema.get("deprecated") and not self.ignore_deprecation:
+                self.errors.metadata_validation_local.update(
+                    {
+                        f"{tsv_path} (as {schema_version.table_schema})": [
+                            "Schema version is deprecated"
+                        ]
+                    }
                 )
                 return
 
-            if schema.get("deprecated") and not self.ignore_deprecation:
-                self.errors.metadata_validation_local[tsv_path].update(
-                    {"Schema version is deprecated": f"{schema_version.table_schema}"}
-                )
-
             local_errors = get_table_errors(tsv_path, schema, report_type)
             if local_errors:
-                self.errors.metadata_validation_local[tsv_path].update(
-                    {f"Local Validation Errors (as {schema_version.table_schema})": local_errors}
+                self.errors.metadata_validation_local.update(
+                    {f"{tsv_path} (as {schema_version.table_schema})": local_errors}
                 )
         else:
             """
@@ -355,10 +401,13 @@ class Upload:
     ):
         url_errors = self._url_checks(tsv_path, schema_name, report_type)
         if url_errors:
-            self.errors.metadata_url_errors[tsv_path].append(url_errors)
-        api_errors = self._api_validation(Path(tsv_path), report_type)
+            self.errors.metadata_url_errors[tsv_path].extend(url_errors)
+        try:
+            api_errors = self._api_validation(Path(tsv_path), report_type)
+        except Exception as e:
+            api_errors = [e]
         if api_errors:
-            self.errors.metadata_validation_api[tsv_path].update(api_errors)
+            self.errors.metadata_validation_api[tsv_path].extend(api_errors)
 
     def _get_reference_errors(self) -> dict:
         errors: Dict[str, Any] = {}
@@ -405,15 +454,15 @@ class Upload:
         self,
         tsv_path: Path,
         report_type: ReportType,
-    ):
-        errors = {}
+    ) -> List[Union[str, Dict]]:
+        errors = []
         response = self._cedar_api_call(tsv_path)
         if response.status_code != 200:
-            errors["Request Errors"] = response.json()
+            raise Exception(response.json())
         elif response.json()["reporting"] and len(response.json()["reporting"]) > 0:
-            errors["Spreadsheet Validator Errors"] = [
-                self._get_message(error, report_type) for error in response.json()["reporting"]
-            ]
+            errors.extend(
+                [self._get_message(error, report_type) for error in response.json()["reporting"]]
+            )
         else:
             logging.info(f"No errors found during CEDAR validation for {tsv_path}!")
         return errors
@@ -537,13 +586,13 @@ class Upload:
 
     def _url_checks(
         self, tsv_path: str, schema_name: str, report_type: ReportType = ReportType.STR
-    ) -> Dict[str, List[Union[str, Dict]]]:
+    ) -> List:
         """
         Check provided UUIDs/HuBMAP IDs for parent_id, sample_id, organ_id.
         Not using get_table_errors because CEDAR schema fields do not match
         the TSV fields, which makes frictionless confused and upset.
         """
-        errors: Dict = {}
+        errors = []
 
         # assay -> parent_sample_id
         # sample -> sample_id
@@ -565,11 +614,11 @@ class Upload:
 
         try:
             url_errors = self._check_matching_urls(tsv_path, constrained_fields, report_type)
-        except Exception as e:
-            errors["URL Errors"] = e
+        except ErrorDictException as e:
+            errors.append(e)
         else:
             if url_errors:
-                errors["URL Errors"] = url_errors
+                errors.extend(url_errors)
         return errors
 
     def _check_matching_urls(
@@ -579,9 +628,11 @@ class Upload:
         fields = rows[0].keys()
         missing_fields = [k for k in constrained_fields.keys() if k not in fields].sort()
         if missing_fields:
-            raise Exception(f"Missing fields: {sorted(missing_fields)}")
+            raise ErrorDictException(f"Missing fields: {sorted(missing_fields)}")
         if not self.globus_token:
-            raise Exception("No token was received to check URL fields against Entity API.")
+            raise ErrorDictException(
+                "No token was received to check URL fields against Entity API."
+            )
         url_errors = []
         for i, row in enumerate(rows):
             check = {k: v for k, v in row.items() if k in constrained_fields}
@@ -648,20 +699,33 @@ class Upload:
         ref: str,
         schema_version: SchemaVersion,
         metadata_path: Union[str, Path],
-    ) -> Optional[Dict]:
+    ) -> Optional[Dict[str, List[str]]]:
         if ref == "data":
             errors = self._check_data_path(schema_version, Path(metadata_path), path_value)
+            if errors:
+                self.errors.directory.update(errors)
         else:
-            errors = self._check_other_path(Path(metadata_path), path_value, ref)
-        return errors
+            other_path = self.directory_path / path_value
+            try:
+                assert other_path.exists()
+            except AssertionError:
+                self.errors.upload_metadata[str(metadata_path)].append(
+                    f"Value '{path_value}' in column '{ref}_path' points to non-existent file: '{self.directory_path / path_value}'"
+                )
+                return
+            try:
+                self._check_other_path(str(other_path))
+            except PreflightError as e:
+                self.errors.upload_metadata[str(metadata_path)].append(
+                    f"Error opening or reading value '{path_value}' from column '{ref}_path': {e.errors}"
+                )
 
     def _get_ref_errors(
         self,
         ref: str,
         schema: SchemaVersion,
         metadata_path: Union[str, Path],
-    ) -> Dict[str, str]:
-        ref_errors: DefaultDict[str, str] = defaultdict(str)
+    ):
         # We don't want to continuously validate shared paths, e.g. contributors.tsv,
         # so this ensures we only check unique paths in a single metadata TSV once
         unique_paths = set()
@@ -679,59 +743,47 @@ class Upload:
                 str(Path(Path(metadata_path).parent, path)) for path in unique_paths
             ]
         for path_value in sorted(unique_paths):
-            ref_error = self._check_path(path_value, ref, schema, metadata_path)
-            if ref_error:
-                ref_errors.update(ref_error)
-        return ref_errors
+            self._check_path(path_value, ref, schema, metadata_path)
 
     def _check_data_path(
         self, schema_version: SchemaVersion, metadata_path: Path, path_value: str
-    ) -> Dict[str, str]:
-        data_path = Path(path_value)
+    ) -> Dict[str, List[str]]:
         errors = {}
+        data_path = Path(path_value)
+        print_path = str(Path(self.directory_path / data_path))
 
         if not schema_version.dir_schema:
             raise Exception(
-                f"No directory schema found for data_path "
-                f"{self.directory_path / data_path} in {metadata_path}!"
+                f"No directory schema found for data_path " f"{print_path} in {metadata_path}!"
             )
 
-        ref_errors = get_data_dir_errors(
-            schema_version.dir_schema,
-            root_path=self.directory_path,
-            data_dir_path=data_path,
-            dataset_ignore_globs=self.dataset_ignore_globs,
-        )
-
-        if ref_errors:
-            errors[f"{str(metadata_path)}, column 'data_path', value '{path_value}'"] = ref_errors
-        return errors
-
-    def _check_other_path(self, metadata_path: Path, other_path_value: str, path_type: str):
-        errors = {}
-        other_path = self.directory_path / other_path_value
         try:
-            schema = get_schema_version(
-                other_path,
-                self.encoding,
-                self.app_context["ingest_url"],
-                self.directory_path,
-                offline=self.offline,
+            ref_errors = get_data_dir_errors(
+                schema_version.dir_schema,
+                root_path=self.directory_path,
+                data_dir_path=data_path,
+                dataset_ignore_globs=self.dataset_ignore_globs,
+            ).popitem()
+            if type(ref_errors[1]) is list:
+                errors[f"{print_path} (as {ref_errors[0]})"] = ref_errors[1]
+            schema_version.dir_schema = ref_errors[0]
+        except FileNotFoundError:
+            self.errors.directory[str(metadata_path)].append(
+                f"Value '{path_value}' in column 'data_path' points to non-existent directory: '{self.directory_path / path_value}'"
             )
         except Exception as e:
-            errors[f"{metadata_path}, column '{path_type}_path', value '{other_path_value}'"] = [e]
-            return errors
-        # TODO: this is broken if you run the tests
-        tsv_ref_errors = self.validation_routine(tsv_paths={str(other_path): schema})
-        # TSV located and read, errors found
-        if tsv_ref_errors and isinstance(tsv_ref_errors, list):
-            errors[other_path] = tsv_ref_errors
-        # Problem with TSV
-        elif tsv_ref_errors and isinstance(tsv_ref_errors, dict):
-            errors[
-                f"{str(metadata_path)}, column '{path_type}_path', value '{other_path_value}'"
-            ] = tsv_ref_errors
+            errors[print_path] = e
         return errors
+
+    def _check_other_path(self, other_path: str):
+        schema = get_schema_version(
+            Path(other_path),
+            self.encoding,
+            self.app_context["ingest_url"],
+            self.directory_path,
+            offline=self.offline,
+        )
+        self.validation_routine(tsv_paths={str(other_path): schema})
 
     def __get_no_ref_errors(self) -> dict:
         referenced_data_paths = (
@@ -817,4 +869,4 @@ class Upload:
                 if col_name in row:
                     reference = f"{tsv_path} (row {i+2})"
                     references[row[col_name]].append(reference)
-        return references
+        return dict(references)